Move seed service in seeds folder
