+++ server/scripts/seeds/index.ts
@@ -3,12 +3,11 @@ import provider from '../../framework/provider';
 import { RequestContext } from '@mikro-orm/core';
 
 const program = new Command('seed');
-program
-  .arguments('[resource]')
-  .action(async (resource = '') => {
-    const { db, seed } = provider.container;
-    await db.connect();
-    return RequestContext.createAsync(db.provider.em, () => seed.run(resource));
-  });
+
+program.action(async () => {
+  const { db, seed } = provider.container;
+  await db.connect();
+  return RequestContext.createAsync(db.provider.em, () => seed.run());
+});
 
 export default program;
